requirements/python-requirements.txt
matplotlib
pillow
numpy
astropy
python-dotenv
